requirements.backup.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiolimiter==1.2.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asgiref==3.8.1
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
beartype==0.20.2
billiard==4.2.1
blinker==1.9.0
blis==1.3.0
blobfile==3.0.0
cachetools==5.5.2
catalogue==2.0.10
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
chess==1.11.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
dacite==1.9.2
datasets==3.5.0
dill==0.3.8
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-encrypted-model-fields==0.6.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docker==7.1.0
evals==3.0.1.post1
evaluate==0.4.3
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.0
griffe==1.6.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
gymnasium==1.1.1
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jiwer==3.1.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.2
kiwisolver==1.4.8
kombu==5.4.2
langchain==0.3.23
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.3.32
language_data==1.3.0
libclang==18.1.1
lxml==5.3.2
lz4==4.4.4
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
mock==5.2.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy==1.15.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.2
numpy==2.1.3
openai==1.68.0
openai-agents==0.0.5
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.7
playwright==1.51.0
pluggy==1.5.0
portalocker==3.1.1
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
sacrebleu==2.5.1
scipy==1.15.2
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
sounddevice==0.5.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy_universal_sentence_encoder==0.4.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
srsly==2.5.1
statsmodels==0.14.4
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
tf_keras==2.19.0
thinc==8.3.6
tiktoken==0.9.0
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250306
types-tqdm==4.67.0.20250417
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
